example/requirements.txt
django>=1.8
django-zipkin
facebook-scribe==2.0
